fix(source_audit): cast numpy floats to Python floats at SQL boundary
Audit log was full of:
    schema "np" does not exist
    LINE 7: ... 0.699202, np.float64(0.817906)...

Root cause: round(numpy.float64, 6) returns numpy.float64 (numpy
overloads __round__ to preserve dtype). psycopg2 has no registered
adapter for numpy.float64 in this venv, so it falls back to str()
which renders as 'np.float64(0.272858)' under the modern numpy repr.
PostgreSQL sees "np.float64" and parses it as a schema-qualified
identifier — "schema np does not exist."

Every comparison row in source_audit.compare_sources hit this and
WARN-failed silently. The full-audit run that caps intelligence_tasks
at 360s was burning most of its time on these failures: scan
completed reporting "171 redundant features, 0 discrepancies" but
zero rows landed in source_accuracy.

Fix: float(round(...)) at the param boundary in source_audit.py:438.
Casts numpy.float64 to Python float, which psycopg2 handles via its
built-in numeric adapter.

This is the actual reason intelligence_tasks was hitting its 360s
budget in cycle 370 — not LLM slowness, it was hammering the DB
with broken INSERTs that all retried and failed.

+++ intelligence/source_audit.py
@@ -439,9 +439,16 @@ def compare_sources(engine: Engine, feature_name: str) -> dict[str, Any]:
                         "fn": feature_name,
                         "sa": sa,
                         "sb": sb,
-                        "corr": round(corr, 6),
-                        "md": round(pct_mean_dev, 6),
-                        "mxd": round(pct_max_dev, 6),
+                        # CAST: round() on a numpy.float64 returns numpy.float64
+                        # (numpy preserves dtype). psycopg2 doesn't have an
+                        # adapter for it and falls back to str() → the literal
+                        # 'np.float64(0.272858)' lands in the SQL, parsing as
+                        # schema-qualified identifier "np.float64" → 'schema
+                        # "np" does not exist' on every row. Force Python
+                        # float at the param boundary.
+                        "corr": float(round(corr, 6)),
+                        "md": float(round(pct_mean_dev, 6)),
+                        "mxd": float(round(pct_max_dev, 6)),
                         "tw": timeliness_winner,
                         "cw": completeness_winner,
                         "aw": accuracy_winner,
